requirements.txt
# Stream Monitor - Dependencies

# Stream capture
streamlink>=6.0.0

# Video/Audio processing
opencv-python>=4.8.0
numpy>=1.24.0
ffmpeg-python>=0.2.0

# Chat clients
twitchio>=2.8.0      # Twitch IRC
pytchat>=0.5.5       # YouTube Live Chat
websockets>=12.0     # Kick (Pusher WebSocket)

# Async support
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
